refactor(logger): \xE2\x99\xBB restore LogEventArgs to InternalLogger
