+++ DomainDetective/DnsPropagationAnalysis.cs
@@ -10,43 +10,6 @@ using System.Threading;
 using System.Threading.Tasks;
 
 namespace DomainDetective {
-    /// <summary>
-    /// Represents a public DNS server used for propagation checks.
-    /// </summary>
-    /// <para>Part of the DomainDetective project.</para>
-    public class PublicDnsEntry {
-        /// <summary>Gets the country of the DNS server.</summary>
-        public string Country { get; init; }
-        /// <summary>Gets the IP address of the DNS server.</summary>
-        public string IPAddress { get; init; }
-        /// <summary>Gets the host name of the DNS server.</summary>
-        public string HostName { get; init; }
-        /// <summary>Gets the location description.</summary>
-        public string Location { get; init; }
-        /// <summary>Gets the ASN of the DNS server.</summary>
-        public string ASN { get; init; }
-        /// <summary>Gets the ASN name of the DNS server.</summary>
-        public string ASNName { get; init; }
-        /// <summary>Gets a value indicating whether the server is enabled.</summary>
-        public bool Enabled { get; init; } = true;
-    }
-
-    /// <summary>
-    /// Result of a DNS propagation query for a single server.
-    /// </summary>
-    /// <para>Part of the DomainDetective project.</para>
-    public class DnsPropagationResult {
-        /// <summary>Gets the server that was queried.</summary>
-        public PublicDnsEntry Server { get; init; }
-        /// <summary>Gets the records returned by the server.</summary>
-        public IEnumerable<string> Records { get; init; }
-        /// <summary>Gets the time the query took.</summary>
-        public TimeSpan Duration { get; init; }
-        /// <summary>Gets a value indicating whether the query succeeded.</summary>
-        public bool Success { get; init; }
-        /// <summary>Gets an error message if the query failed.</summary>
-        public string Error { get; init; }
-    }
 
     /// <summary>
     /// Provides DNS propagation checks across many public servers.
