+++ DomainDetective/DnsblConfiguration.cs
@@ -1,29 +1,15 @@
-namespace DomainDetective;
-
-using System;
+namespace DomainDetective {
 using System.Collections.Generic;
 
-/// <summary>
-/// Configuration for DNS block list providers.
-/// </summary>
-/// <para>Part of the DomainDetective project.</para>
-public class DnsblConfiguration {
     /// <summary>
-    /// Gets or sets the list of DNSBL providers.
+    /// Configuration for DNS block list providers.
     /// </summary>
-    public List<DnsblEntry> Providers { get; set; } = new();
+    /// <para>Part of the DomainDetective project.</para>
+    public class DnsblConfiguration {
+        /// <summary>Gets or sets the list of DNSBL providers.</summary>
+        public List<DnsblEntry> Providers { get; set; } = new();
 
-    /// <summary>
-    /// Gets or sets domain based block lists.
-    /// </summary>
-    public List<DnsblEntry> DomainBlockLists { get; set; } = new();
-
-}
-
-/// <summary>
-/// Provider specific reply code configuration.
-/// </summary>
-public class DnsblReplyCode {
-    public bool IsListed { get; set; }
-    public string Meaning { get; set; }
+        /// <summary>Gets or sets domain based block lists.</summary>
+        public List<DnsblEntry> DomainBlockLists { get; set; } = new();
+    }
 }
